werkend krijgen mvc op origineel programma
state nog niet helemaal duidelijk, maar het werkt zonder states
