+++ src/IObserver.java
@@ -1,5 +1,3 @@
-
 public interface IObserver {
-
-	public void update(ISubject subject);
+	public void update(Presentation subject, Slide data);
 }
