+++ src/IPresentatieModel.java
@@ -1,4 +1,6 @@
-public interface ISubject {
+public interface IPresentatieModel extends ISubject {
+	public int getState();
+	public void ChangeState();
 	public void addObserver(IObserver observer);
 	public void removeObserver(IObserver observer);
 	public void notifyAllObservers();
