+++ src/SlideViewerFrame.java
@@ -14,8 +14,8 @@ import javax.swing.JFrame;
  * @version 1.6 2014/05/16 Sylvia Stuurman
 */
 
-public abstract class SlideViewerFrame extends JFrame implements IObserver {
-	//observer dingen
+public class SlideViewerFrame extends JFrame implements IObserver {
+	//observer dingen TODO
 	// private int observerState = subject->GetState();
 	
 	private static final long serialVersionUID = 3227L;
@@ -24,11 +24,13 @@ public abstract class SlideViewerFrame extends JFrame implements IObserver {
 	public final static int WIDTH = 1200;
 	public final static int HEIGHT = 800;
 	
+	private SlideViewerComponent component;
+	
 	public SlideViewerFrame(String title, Presentation presentation) {
 		super(title);
-		SlideViewerComponent slideViewerComponent = new SlideViewerComponent(presentation, this);
-		presentation.setShowView(slideViewerComponent);
-		setupWindow(slideViewerComponent, presentation);
+		this.component = new SlideViewerComponent(presentation, this);
+		presentation.setShowView(this.component);
+		setupWindow(this.component, presentation);
 	}
 
 // De GUI opzetten
@@ -47,8 +49,9 @@ public abstract class SlideViewerFrame extends JFrame implements IObserver {
 		setVisible(true);
 	}
 
-	public void update(ISubject subject) {
+	public void update(Presentation subject, Slide data) {
 		// krijgt het model mee
-		// voert de update uit door dit weer door te geven naar viewerComponent		
+		// voert de update uit door dit weer door te geven naar viewerComponent	
+		this.component.update(subject, data);
 	}
 }
